Add a script to convert Polaris-output .MRC file to chunked marcxml, using yaz-marcdump.
This works in two phases.  First it uses yaz-marcdump to split the
single large file into chunks.  (It would be possible to avoid this
step by manually having Polaris output the chunks, but this way is
more automated and therefore less error-prone, not to mention less
tedious.)  Then as a second step it uses yaz-marcdump to convert
each of those chunks into MARCXML format.

I'm not sure I understand why yaz-marcdump can't do both things as a
single step, but whatever, this works.